+++ node/src/Commands.ts
@@ -1649,3 +1649,25 @@ export function createLolwut(options?: LolwutOptions): command_request.Command {
 
     return createCommand(RequestType.Lolwut, args);
 }
+
+export enum FlushMode {
+    /**
+     * Flushes synchronously.
+     *
+     * @since Valkey 6.2 and above.
+     */
+    SYNC,
+    /** Flushes asynchronously. */
+    ASYNC,
+}
+
+/**
+ * @internal
+ */
+export function createFlushAll(mode?: FlushMode): command_request.Command {
+    if (mode) {
+        return createCommand(RequestType.FlushAll, [mode.toString()]);
+    } else {
+        return createCommand(RequestType.FlushAll, []);
+    }
+}