+++ node/src/GlideClient.ts
@@ -5,6 +5,7 @@
 import * as net from "net";
 import { BaseClient, BaseClientConfiguration, ReturnType } from "./BaseClient";
 import {
+    FlushMode,
     InfoOptions,
     LolwutOptions,
     createClientGetName,
@@ -15,6 +16,7 @@ import {
     createConfigSet,
     createCustomCommand,
     createEcho,
+    createFlushAll,
     createInfo,
     createLolwut,
     createPing,
@@ -329,5 +331,28 @@ export class GlideClient extends BaseClient {
      */
     public lolwut(options?: LolwutOptions): Promise<string> {
         return this.createWritePromise(createLolwut(options));
+    }
+
+    /**
+     * Deletes all the keys of all the existing databases. This command never fails.
+     * The command will be routed to all primary nodes.
+     *
+     * See https://valkey.io/commands/flushall/ for more details.
+     *
+     * @param mode The flushing mode, could be either {@link FlushMode#SYNC} or {@link FlushMode#ASYNC}.
+     * @returns `OK`.
+     *
+     * @example
+     * ```typescript
+     * const result = await client.flushall(FlushMode.SYNC);
+     * console.log(result); // Output: 'OK'
+     * ```
+     */
+    public flushall(mode?: FlushMode): Promise<string> {
+        if (mode) {
+            return this.createWritePromise(createFlushAll(mode));
+        } else {
+            return this.createWritePromise(createFlushAll());
+        }
     }
 }