+++ node/tests/TestUtilities.ts
@@ -10,6 +10,7 @@ import {
     BaseClient,
     BaseClientConfiguration,
     ClusterTransaction,
+    FlushMode,
     GlideClient,
     GlideClusterClient,
     InsertPosition,
@@ -313,6 +314,8 @@ export async function transactionTest(
     const field = uuidv4();
     const value = uuidv4();
     const args: ReturnType[] = [];
+    baseTransaction.flushall();
+    args.push("OK");
     baseTransaction.set(key1, "bar");
     args.push("OK");
     baseTransaction.objectEncoding(key1);